+++ src/constants/index.js
@@ -3,3 +3,4 @@ export * from './rawSearchableIndexes';
 export * from './searchableIndexesMap';
 export * from './searchResultsListColumns';
 export * from './filterConfig';
+export * from './sortOrders';