+++ src/constants/sortOrders.js
@@ -1,0 +1,4 @@
+export const sortOrders = {
+  ASC: 'ascending',
+  DES: 'descending',
+};