+++ src/hooks/index.js
@@ -1,1 +1,2 @@
+export { default as useSortColumnManager } from './useSortColumnManager';
 export * from './useSectionToggle';